requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph & LangChain
langgraph>=0.0.26
langchain>=0.1.6
langchain-openai>=0.0.5
langchain-community>=0.0.19

# Vector database
chromadb>=0.4.0

# LlamaIndex GraphRAG
llama-index-core>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-graph-stores-nebula>=0.2.0
nebula3-python>=3.8.0

# OpenAI API
openai>=1.10.0

# Frontend
streamlit>=1.31.0
requests>=2.31.0
python-multipart>=0.0.9

# Speech recognition
dashscope>=1.25.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
